requirements.txt
# Data Processing
pandas==2.1.3
numpy==1.24.3

# Machine Learning Models
scikit-learn==1.3.2
xgboost==2.0.3
statsmodels==0.14.0
prophet==1.1.5
tensorflow==2.14.0

# Data Fetching
yfinance==0.2.32
ccxt==4.0.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
joblib==1.3.2
tqdm==4.66.1